// apps/admin-dashboard/src/components/common/AdminBreadcrumb.d.ts
import React from 'react';
interface BreadcrumbItem {
    label: string;
    path?: string;
}
interface AdminBreadcrumbProps {
    items?: BreadcrumbItem[];
    className?: string;
}
declare const AdminBreadcrumb: React.FC<AdminBreadcrumbProps>;
export default AdminBreadcrumb;
//# sourceMappingURL=AdminBreadcrumb.d.ts.map